move all repo list functions as a standalone file
